Working UI
There is communication between the PLC and Python. The UI contains all 3 modes, a countdown and a pretension button with a pretension spinbox. All the reflecting modes have an "input" table, which if enabled, will be the input line for the Live Graph. The Live graph detects the torque sensor values and recreates a sensor line. There are preset values for each mode.